Adds back the gamer gear enclave LARP
haha funni you cant use this BoS because it's against your honor code or whatever. *inhales a drag* mmmmm pubbie tears.

also removes dumb reference /properly/ this time. it's gonna stay as a bus item because there are tons of other non-FO13 references in the code from the Home Run Bat to Holoparasites. It's not that bad of an item either and I personally like the sprite.