Water Update
The Water Update adds a plecera of Water-Bottles, and Nuka-Colas to the map.

There is also a Water-Hoarder group of raiders who wish to hoard water for there own profit.

The factions also have some flasks or canteens to store some water on the go.

Kebabtown's Richer Settlers have atleast a single nuka-cola in every fridge!

The bar of Kebabtown also has a new Water-Cooler.

Due to the recent sky-rocket in thirsty wanderers, Empty Plastic Waterbottles have began trashing the area!